Update to work with new CompletableFutures system

+++ src/main/java/net/factmc/FactBukkit/gui/ServerGUI.java
@@ -109,7 +109,7 @@ public class ServerGUI implements Listener {
 				}
 				
 			}
-		}, (long) 20);
+		}, 20L);
 		
 	}
